Move PncUpdateDatasetParsedXml types into root types folder
